Resolve all project and todo creation forms
