+++ src/index.js
@@ -1,96 +1,8 @@
 import Project, { showProjects, projectExists} from './project.js';
 import Todo, { showTodos } from './todo.js';
 
-let allProjects = [new Project('default')];
-let currentPrj = allProjects[0];
-const todoSection = document.getElementById('project-todos');
-const projectTodos = document.getElementById('todo-creation-form');
-
-const getProjectsFromLocalStorage = () => {
-  const localStorageProjects = localStorage.getItem('allProjects');
-  if (localStorageProjects === null) {
-    allProjects = [new Project('default')];
-    localStorage.setItem('allProjects', JSON.stringify(allProjects));
-  } else {
-    allProjects = JSON.parse(localStorageProjects);
-  }
-  return allProjects;
-};
-
-const onProjectClick = (e, newTodoCreationForm) => {
-  const projectDiv = e.target;
-  const index = Number(projectDiv.dataset.index);
-  allProjects = getProjectsFromLocalStorage();
-  currentPrj = allProjects[index];
-  todoSection.innerHTML = '';
-  showTodos(currentPrj, newTodoCreationForm);
-  newTodoCreationForm(showTodos);
-  return projectDiv;
-};
-
-const newTodoCreationForm = (showTodos) => {
-  const newTodoForm = document.createElement('form');
-  newTodoForm.id = 'new-todo-creation-form';
-  const todoTitle = document.createElement('input');
-  todoTitle.setAttribute('type', 'text');
-  todoTitle.setAttribute('placeholder', 'todo title');
-  todoTitle.id = 'todo-title';
-
-  const todoDesc = document.createElement('input');
-  todoDesc.setAttribute('type', 'text');
-  todoDesc.setAttribute('placeholder', 'some description..');
-  todoDesc.id = 'todo-desc';
-
-  const todoDueDate = document.createElement('input');
-  todoDueDate.setAttribute('type', 'date');
-  todoDueDate.setAttribute('placeholder', 'due date');
-  todoDueDate.id = 'todo-priority';
-
-  const todoPriority = document.createElement('select');
-  todoPriority.setAttribute('id', 'priority');
-  todoPriority.id = 'todo-dueDate';
-
-  const todoPriorities = ['Very high', 'High', 'Moderate', 'Low', 'Useless'];
-  const options = todoPriorities.map((priority) => {
-    const value = priority.toLowerCase();
-    return `<option value="${value}">${priority}</option>`;
-  });
-  todoPriority.innerHTML = options;
-
-  const todoSubmit = document.createElement('input');
-  todoSubmit.setAttribute('type', 'submit');
-  todoSubmit.id = 'todo-form-submit';
-
-  newTodoForm.append(todoTitle, todoDesc, todoDueDate, todoPriority, todoSubmit);
-  projectTodos.innerHTML = '';
-  projectTodos.append(newTodoForm);
-
-  newTodoForm.addEventListener('submit', (e) => {
-    e.preventDefault();
-    const title = document.getElementById('todo-title').value;
-    const description = document.getElementById('todo-desc').value;
-    const priority = document.getElementById('todo-priority').value;
-    const dueDate = document.getElementById('todo-dueDate').value;
-    const newTodoValues = new Todo(title, description, priority, dueDate);
-    allProjects = getProjectsFromLocalStorage();
-    const currentPrjIdx = allProjects.findIndex((project) => project.title === currentPrj.title);
-    allProjects[currentPrjIdx].todos.push(newTodoValues);
-    localStorage.setItem('allProjects', JSON.stringify(allProjects));
-    showTodos(currentPrj, newTodoCreationForm);
-  });
-  return newTodoForm;
-};
-
-const projectCard = (project, idx) => {
-  const projectDiv = document.createElement('a');
-  projectDiv.id = 'project-div';
-  projectDiv.textContent = project.title;
-  projectDiv.dataset.index = idx;
-  projectDiv.addEventListener('click', (e) => {
-    onProjectClick(e, newTodoCreationForm);
-  });
-  return projectDiv;
-};
+export const todoSection = document.getElementById('project-todos');
+export const projectTodos = document.getElementById('todo-creation-form');
 
 const addProjectToLocalStorage = (newProject) => {
   const allProjects = getProjectsFromLocalStorage();
